Move capabilities handling to own file and refactor http client pool.
